Change Login and Register pages into components, adjust files structure
