+++ RecipeManager.Shared.Contracts/User/Registration/PasswordValidator.cs
@@ -2,18 +2,6 @@
 
 namespace RecipeManager.Shared.Contracts.User.Registration;
 
-public sealed class UserNameValidator : AbstractValidator<string>
-{
-    public UserNameValidator()
-    {
-        RuleFor(userName => userName)
-            .NotEmpty()
-                .WithMessage("Username is required")
-            .MaximumLength(15)
-                .WithMessage("Username must not exceed 15 characters");
-    }
-}
-
 public sealed class PasswordValidator : AbstractValidator<string>
 {
     public PasswordValidator()
@@ -33,15 +21,3 @@ public sealed class PasswordValidator : AbstractValidator<string>
                 .WithMessage("Password must contain at least one special character");
     }
 }
-
-public sealed class EmailValidator : AbstractValidator<string>
-{
-    public EmailValidator()
-    {
-        RuleFor(email => email)
-            .NotEmpty()
-                .WithMessage("Email is required")
-            .EmailAddress()
-                .WithMessage("Invalid email format");
-    }
-}
