+++ RecipeManager.UI.Blazor/RecipeManager.UI.Blazor/Features/Account/Login/LoginForm.razor
@@ -1,6 +1,4 @@
-﻿@page "/login"
-
-@using FluentValidation
+﻿@using FluentValidation
 @using MudBlazor
 @using RecipeManager.Shared.Contracts.User.Registration
 @using RecipeManager.UI.Blazor.Components.Extensions
@@ -11,7 +9,7 @@
 @inject NavigationManager Navigation
 @inject IdentityApiService IdentityApi
 
-<MudPaper Class="pa-6 mx-auto" MaxWidth="400px">
+@* <MudPaper Class="pa-6 mx-auto" MaxWidth="400px"> *@
     <MudForm Model="@loginModel" ValidationDelay="0" @bind-IsValid="@isSubmitEnabled">
         <MudText Typo="Typo.h5" Class="mb-4">Sign in</MudText>
 
@@ -23,16 +21,16 @@
         </MudButton>
         <MudText ReadOnly="true" Color="Color.Error">@errorMessage</MudText>
 
-        <MudGrid AlignItems="AlignItems.Center" Class="mt-6">
-            <MudItem xs="12" Class="d-flex justify-space-between align-center mt-4">
-                <MudText Typo="Typo.body1">Not a member yet?</MudText>
-                <MudButton Variant="Variant.Text" Color="Color.Primary" OnClick="NavigateToRegisterPage">
-                    Register
-                </MudButton>
-            </MudItem>
-        </MudGrid>
+        @* <MudGrid AlignItems="AlignItems.Center" Class="mt-6">
+        @*     <MudItem xs="12" Class="d-flex justify-space-between align-center mt-4"> *@
+        @*         <MudText Typo="Typo.body1">Not a member yet?</MudText> *@
+        @*         <MudButton Variant="Variant.Text" Color="Color.Primary" OnClick="NavigateToRegisterPage"> *@
+        @*             Register *@
+        @*         </MudButton> *@
+        @*     </MudItem> *@
+        @* </MudGrid> *@
     </MudForm>
-</MudPaper>
+ @* </MudPaper> *@
 
 @code {
     private LoginRequest loginModel = new();
