+++ RecipeManager.UI.Blazor/RecipeManager.UI.Blazor/Features/Account/Register/RegisterForm.razor
@@ -1,0 +1,40 @@
+﻿@using FluentValidation
+@using MudBlazor
+@using RecipeManager.Shared.Contracts.User.Registration
+@using RecipeManager.UI.Blazor.Components.Extensions
+@using RecipeManager.UI.Blazor.Services
+
+@inject ISnackbar Snackbar
+@inject NavigationManager Navigation
+@inject IdentityApiService IdentityApi
+
+<MudForm Model="@registrationModel" Validation="@(validator.ValidateValue)" ValidationDelay="0" @bind-IsValid="@isSubmitEnabled">
+    <MudText Typo="Typo.h5" Class="mb-4">Register</MudText>
+
+    <MudTextField @bind-Value="registrationModel.UserName" Label="Username" For="@(() => registrationModel.UserName)" Required=true/>
+    <MudTextField @bind-Value="registrationModel.Email" Label="Email" For="@(() => registrationModel.Email)"  Required=true/>
+    <MudTextField @bind-Value="registrationModel.Password" Label="Password" For="@(() => registrationModel.Password)" InputType="InputType.Password" Required=true />
+    <MudTextField @bind-Value="registrationModel.ConfirmationPassword" Label="Confirmation Password" For="@(() => registrationModel.ConfirmationPassword)" InputType="InputType.Password" Required=true />
+
+    <MudButton Variant="Variant.Filled" Color="Color.Primary" OnClick="OnSubmit" Disabled="@(!isSubmitEnabled)" Class="mt-4" FullWidth>
+        Register
+    </MudButton>
+    <MudText ReadOnly="true" Color="Color.Error">@errorMessage</MudText>
+</MudForm>
+
+@code {
+    private UserRegistrationModel registrationModel = new();
+    private readonly UserRegistrationValidator validator = new();
+    private bool isSubmitEnabled = false;
+    private string errorMessage = "";
+
+    private async Task OnSubmit()
+    {
+        await IdentityApi.RegisterUser(registrationModel);
+    }
+
+    private void NavigateToLoginPage()
+    {
+        Navigation.NavigateTo("/login");
+    }
+}
